Move x86-64 pKVM PCI CAM to submodule.
